dashboard/api/requirements.txt
fastapi
uvicorn
redis
psycopg2-binary
pydantic
python-dotenv
